Error Fix on Data Collector
When cleaning up the code for final submission to APPS, a typo was made in the code that caused an incorrect calculation under certain circumstances. This error did NOT affect the initial analysis with either Geranium maculatum or Linanthus parryae. The error occurred after analysis (and was verified by reviewing the data and original files) when adding comments and other minor cleanups to make the code more readable.